Address bug reported in issue 60
As part of issue 60 it was reported that a vROps alertName may be an
empty string. If this is the case then notifications to pagerduty will
fail as pagerduty requires that the the description exist and not be the
empty string. Change for and catch this condition.

Temporarily disable opsgenie tests until switched to v2 API.